Split protected views
