+++ statistics/templates/statistics/base.html
@@ -69,7 +69,7 @@
       {% if not no_footer %}
       <hr>
       <footer>
-        <p style="float: left; width: 33%; text-align: left;">&copy; European Youth Parliament 2016</p>
+        <p style="float: left; width: 33%; text-align: left;">&copy; European Youth Parliament 2017</p>
         <p style="float: left; width: 33%; text-align: center;">Made by Oliver Stenbom and Tom Wolfskämpf</p>
         <p style="float: left; width: 33%; text-align: end;"><a target="_blank" href="https://gitreports.com/issue/eyp-developers/statistics">Bugs and Feature Requests</a> | <a href="{% url 'statistics:changelog' %}">Changelog</a></p>
       </footer>
