+++ statistics/urls.py
@@ -21,7 +21,7 @@ urlpatterns = [
     url(r'^highscores/$', high_scores, name='high_scores'),
     url(r'^overview/(?P<session_id>[0-9]+)/$', overview, name='overview'),
     url(r'^edit/(?P<session_id>[0-9]+)/$', edit, name='edit'),
-    url(r'^committees/(?P<session_id>[0-9]+)/$', add, name='add'),
+    url(r'^committees/(?P<session_id>[0-9]+)/$', create_committee, name='create_committee'),
     url(r'^manage/(?P<session_id>[0-9]+)/$', manage, name='manage'),
     url(r'^gender/(?P<session_id>[0-9]+)/$', gender, name='gender'),
     url(r'^runningorder/(?P<session_id>[0-9]+)/$', runningorder, name='runningorder'),
